Support gzip & deflate response encodings
Adds support for gzip and deflate response encodings for large response bodies. This will allow the client to use less bandwidth and receive the responses faster when paginating through large lists. The Recurly responses automatically switch to a gzip response type when the body is at least a few kilobytes. Now the client can support plain-text and gzip'ed responses.