Support request types as bodies
Previously, the body parameter of the operations was just defined as a plain object.
This uses their proper Request types as we do in the other statically typed clients.
We're not considering this a breaking change because if you had passed an invalid object before, the server would already be stopping you